requirements.txt
# ─────────────────────────────────────────────
# Multi-Modal Diabetic Complication Predictor
# requirements.txt
# Install with: pip install -r requirements.txt
# ─────────────────────────────────────────────

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0

# Pretrained Image Models
timm>=0.9.12

# Data & Numerics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Explainability
shap>=0.43.0
grad-cam>=1.4.8          # pytorch-grad-cam

# Hyperparameter Optimization
optuna>=3.4.0

# Experiment Tracking
wandb>=0.16.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# API & Deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.6.0
